Make common component GenericComponent to reduce code reuse
