+++ client/src/Book.tsx
@@ -2,12 +2,7 @@
 // Licensed under the MIT license.
 
 import { LayerHost, Spinner, Stack, ThemeProvider } from '@fluentui/react';
-import {
-  backgroundStyles,
-  fullSizeStyles,
-  getDefaultContainerStyles,
-  getDefaultLayerHostStyles
-} from './styles/Common.styles';
+import { backgroundStyles, fullSizeStyles } from './styles/Common.styles';
 import { Header } from './Header';
 import './styles/Common.css';
 import { fetchConfig } from './utils/FetchConfig';
@@ -16,7 +11,7 @@ import { GenericError } from './components/GenericError';
 import { useEffect, useState } from 'react';
 import { BookingsPage } from './components/book/BookingsPage';
 import { NoSchedulingPage } from './components/book/NoSchedulingPage';
-import MobileDetect from 'mobile-detect';
+import GenericContainer from './components/GenericContainer';
 
 export const Book = (): JSX.Element => {
   const [config, setConfig] = useState<AppConfigModel | undefined>(undefined);
@@ -39,11 +34,6 @@ export const Book = (): JSX.Element => {
   }, []);
 
   if (config) {
-    const containerStyles = getDefaultContainerStyles(
-      config.theme,
-      new MobileDetect(window.navigator.userAgent).mobile()
-    );
-
     return (
       <ThemeProvider theme={config.theme} style={{ height: '100%' }}>
         <Stack styles={backgroundStyles(config.theme)}>
@@ -59,11 +49,9 @@ export const Book = (): JSX.Element => {
               <BookingsPage config={config} />
             </LayerHost>
           ) : (
-            <LayerHost id={PARENT_ID} style={getDefaultLayerHostStyles()}>
-              <Stack styles={containerStyles} tokens={{ childrenGap: 15 }}>
-                <NoSchedulingPage />
-              </Stack>
-            </LayerHost>
+            <GenericContainer layerHostId={PARENT_ID} theme={config.theme}>
+              <NoSchedulingPage />
+            </GenericContainer>
           )}
         </Stack>
       </ThemeProvider>
