+++ client/src/Home.test.tsx
@@ -2,7 +2,6 @@
 // Licensed under the MIT license.
 
 import { mount } from 'enzyme';
-import renderer from 'react-test-renderer';
 import * as FetchConfig from './utils/FetchConfig';
 import * as CreateRoom from './utils/FetchRoomsResponse';
 import { runFakeTimers } from './utils/TestUtils';
@@ -12,12 +11,6 @@ import { Home } from './Home';
 import { AppConfigModel } from './models/ConfigModel';
 import { GenericError } from './components/GenericError';
 import { CreateRoomResponse } from './models/RoomModel';
-
-let userAgentGetter: any = undefined;
-
-beforeEach(() => {
-  userAgentGetter = jest.spyOn(window.navigator, 'userAgent', 'get');
-});
 
 describe('Home', () => {
   it('should render loading spinner when config is not loaded', async () => {
@@ -98,16 +91,5 @@ describe('Home', () => {
 
     expect(spinners.length).toBe(0);
     expect(genericErrors.length).toBe(1);
-  });
-
-  it.each([['desktop'], ['mobile']])('matches snapshot where form factor is %s', async (formFactor: string) => {
-    if (formFactor === 'mobile') {
-      const mobileSafariUserAgent =
-        'Mozilla/5.0 (iPhone; CPU iPhone OS 14_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/14.0.3 Mobile/15E148 Safari/604.1';
-      userAgentGetter.mockReturnValue(mobileSafariUserAgent);
-    }
-
-    const component = renderer.create(<Home />).toJSON();
-    expect(component).toMatchSnapshot();
   });
 });
