+++ client/src/Home.tsx
@@ -1,13 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { LayerHost, Spinner, Stack, ThemeProvider } from '@fluentui/react';
-import {
-  backgroundStyles,
-  fullSizeStyles,
-  getDefaultContainerStyles,
-  getDefaultLayerHostStyles
-} from './styles/Common.styles';
+import { Spinner, Stack, ThemeProvider } from '@fluentui/react';
+import { backgroundStyles, fullSizeStyles } from './styles/Common.styles';
 import { useEffect, useState } from 'react';
 import { fetchConfig } from './utils/FetchConfig';
 import { GenericError } from './components/GenericError';
@@ -15,7 +10,7 @@ import { AppConfigModel } from './models/ConfigModel';
 import { HomeComponent } from './components/home/HomeComponent';
 import { Header } from './Header';
 import './styles/Common.css';
-import MobileDetect from 'mobile-detect';
+import GenericContainer from './components/GenericContainer';
 
 const PARENT_ID = 'HomeSection';
 
@@ -45,24 +40,17 @@ export const Home = (): JSX.Element => {
     return <Spinner styles={fullSizeStyles} />;
   }
 
-  const containerStyles = getDefaultContainerStyles(
-    config.theme,
-    new MobileDetect(window.navigator.userAgent).mobile()
-  );
-
   return (
     <ThemeProvider theme={config.theme} style={{ height: '100%' }}>
       <Stack styles={backgroundStyles(config.theme)}>
         <Header companyName={config.companyName} parentid={PARENT_ID} />
-        <LayerHost id={PARENT_ID} style={getDefaultLayerHostStyles()}>
-          <Stack styles={containerStyles} tokens={{ childrenGap: 15 }}>
-            <HomeComponent
-              companyName={config.companyName}
-              theme={config.theme}
-              onDisplayError={(error) => setError(error)}
-            />
-          </Stack>
-        </LayerHost>
+        <GenericContainer layerHostId={PARENT_ID} theme={config.theme}>
+          <HomeComponent
+            companyName={config.companyName}
+            theme={config.theme}
+            onDisplayError={(error) => setError(error)}
+          />
+        </GenericContainer>
       </Stack>
     </ThemeProvider>
   );
